test(llm): move the finish-reason conformance suite out of the rig adapter
`rig_adapter.rs` had grown to 4,580 lines of mixed production and test code,
and this PR added roughly 850 more — a matrix spanning hundreds of lines of
provider payloads. Following the adapter's conversion logic meant scrolling
past all of it.

The provider fixtures, payload builders, streaming script model and the two
matrices now live in `src/rig_adapter/tests/finish_reason_tests.rs`, a child
of the adapter's `tests` module, so `super::*` still reaches both the
adapter's items and the shared test helpers. The adapter file drops to 3,951
lines. No test content changed; the whole suite still runs.
